requirements.txt
pandas>=1.3.0
numpy>=1.19.5
scikit-learn>=0.24.2
nltk>=3.6.3
xgboost>=1.5.0
pyyaml>=5.4.1 